requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.7.3
asgiref==3.4.1
astroid==2.8.0
attrs==21.2.0
autopep8==1.5.7
bandit==1.7.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
coverage==5.5
cryptography==35.0.0
dnspython==2.1.0
ecdsa==0.17.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.1.3
fastapi==0.68.1
flake8==3.9.2
gitdb==4.0.7
gitpython==3.1.24
greenlet==1.1.1
h11==0.12.0
idna==3.2
infinity==1.5
iniconfig==1.1.1
interval==1.0.0
intervals==0.9.2
isort==5.9.3
lazy-object-proxy==1.6.0
mako==1.1.5
markupsafe==2.0.1
mccabe==0.6.1
packaging==21.0; python_version >= '3.6'
passlib[bcrypt]==1.7.4
pbr==5.6.0
platformdirs==2.3.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.8.2
pyflakes==2.3.1
pylint==2.11.1
pyparsing==2.4.7
pytest-cov==2.12.1
pytest-lazy-fixture==0.6.3
pytest-mock==3.6.1
pytest==6.2.5
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
pyyaml==5.4.1
requests==2.26.0
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
six==1.16.0
smmap==4.0.0
sqlalchemy==1.4.23
starlette==0.14.2
stevedore==3.4.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
wrapt==1.12.1
